+++ contracts/Oracle/WethToWethOracle.sol
@@ -6,7 +6,6 @@ import "./ICryptoPairOracle.sol";
 // Fixed window oracle that recomputes the average price for the entire period once every period
 // Note that the price average is only guaranteed to be over at least 1 period, but may be over a longer period
 contract WethToWethOracle is ICryptoPairOracle {
-    
     address internal wethAddress;
 
     constructor(address _wethAddress) public {
@@ -15,7 +14,7 @@ contract WethToWethOracle is ICryptoPairOracle {
         wethAddress = _wethAddress;
     }
 
-    function consult(address tokenIn, uint256 amountIn) external view override returns (uint amountOut) {
+    function consult(address tokenIn, uint256 amountIn) external view override returns (uint256 amountOut) {
         require(tokenIn == wethAddress, "This oracle only accepts consulting WETH input");
 
         return amountIn;
